mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.16.1",
    "mcp>=1.9.0",
    "openai>=2.15.0",
    "python-dotenv>=1.1.0",
    "typing-extensions>=4.13.2",
]
